# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mock FPS API",
    "description": "Mock server simulating the Form3 Faster Payments Service (FPS) API. Returns realistic JSON:API responses with async lifecycle status transitions, stand-in mode, and webhook notifications.",
    "version": "0.1.2",
    "contact": {
      "name": "hum3",
      "url": "https://codeberg.org/hum3/mock-fps"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
...
  "paths": {
...
    "/v1/transaction/payments/{paymentID}": {
      "get": {
        "tags": ["Payments"],
        "summary": "Get payment",
        "operationId": "getPayment",
        "parameters": [
          { "$ref": "#/components/parameters/paymentID" }
        ],
        "responses": {
          "200": {
            "description": "Payment found",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDataEnvelope"
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/v1/transaction/payments/{paymentID}/submissions": {
      "post": {
        "tags": ["Payment Submissions"],
        "summary": "Create submission",
        "description": "Creates a payment submission and triggers async lifecycle: accepted -> validation_pending -> limit_check_pending -> limit_check_passed -> released_to_gateway -> queued_for_delivery -> submitted -> delivery_confirmed",
        "operationId": "createPaymentSubmission",
        "parameters": [
          { "$ref": "#/components/parameters/paymentID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/vnd.api+json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentSubmissionDataEnvelope"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Submission created",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentSubmissionDataEnvelope"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "415": { "$ref": "#/components/responses/UnsupportedMediaType" }
        }
      }
    },
    "/v1/transaction/payments/{paymentID}/submissions/{submissionID}": {
      "get": {
        "tags": ["Payment Submissions"],
        "summary": "Get submission",
        "operationId": "getPaymentSubmission",
        "parameters": [
          { "$ref": "#/components/parameters/paymentID" },
          { "$ref": "#/components/parameters/submissionID" }
        ],
        "responses": {
          "200": {
            "description": "Submission found",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentSubmissionDataEnvelope"
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/v1/transaction/payments/{paymentID}/admissions": {
      "post": {
        "tags": ["Payment Admissions"],
        "summary": "Create admission",
        "description": "Creates a payment admission and triggers async lifecycle: pending -> confirmed",
        "operationId": "createPaymentAdmission",
        "parameters": [
          { "$ref": "#/components/parameters/paymentID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/vnd.api+json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentAdmissionDataEnvelope"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Admission created",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentAdmissionDataEnvelope"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "415": { "$ref": "#/components/responses/UnsupportedMediaType" }
        }
      }
    },
    "/v1/transaction/payments/{paymentID}/admissions/{admissionID}": {
      "get": {
        "tags": ["Payment Admissions"],
        "summary": "Get admission",
        "operationId": "getPaymentAdmission",
        "parameters": [
          { "$ref": "#/components/parameters/paymentID" },
          { "$ref": "#/components/parameters/admissionID" }
        ],
        "responses": {
          "200": {
            "description": "Admission found",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentAdmissionDataEnvelope"
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/v1/transaction/payments/{paymentID}/admissions/{admissionID}/tasks/{taskID}": {
      "patch": {
        "tags": ["Admission Tasks"],
        "summary": "Update admission task",
        "operationId": "updateAdmissionTask",
        "parameters": [
          { "$ref": "#/components/parameters/paymentID" },
          { "$ref": "#/components/parameters/admissionID" },
          { "$ref": "#/components/parameters/taskID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/vnd.api+json": {
              "schema": {
                "$ref": "#/components/schemas/AdmissionTaskDataEnvelope"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmissionTaskDataEnvelope"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "415": { "$ref": "#/components/responses/UnsupportedMediaType" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "paymentID": {
        "name": "paymentID",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
      "submissionID": {
        "name": "submissionID",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
      "admissionID": {
        "name": "admissionID",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
      "taskID": {
        "name": "taskID",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/vnd.api+json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/vnd.api+json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Conflict": {
        "description": "Resource already exists",
        "content": {
          "application/vnd.api+json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "UnsupportedMediaType": {
        "description": "Content-Type must be application/vnd.api+json",
        "content": {
          "application/vnd.api+json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
      "Resource": {
        "type": "object",
        "properties": {
          "type": { "type": "string" },
          "id": { "type": "string", "format": "uuid" },
          "organisation_id": { "type": "string" },
          "version": { "type": "integer" },
          "created_on": { "type": "string", "format": "date-time" },
          "modified_on": { "type": "string", "format": "date-time" }
        },
        "required": ["type", "id"]
      },
      "AccountParty": {
        "type": "object",
        "properties": {
          "account_name": { "type": "string" },
          "account_number": { "type": "string" },
          "sort_code": { "type": "string" }
        }
      },
      "ChargesInformation": {
        "type": "object",
        "properties": {
          "bearer_code": { "type": "string" }
        }
      },
      "FxInfo": {
        "type": "object",
        "properties": {
          "contract_reference": { "type": "string" },
          "exchange_rate": { "type": "string" },
          "original_amount": { "type": "string" },
          "original_currency": { "type": "string" }
        }
      },
      "RelationshipData": {
        "type": "object",
        "properties": {
          "type": { "type": "string" },
          "id": { "type": "string", "format": "uuid" }
        },
        "required": ["type", "id"]
      },
      "Relationship": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/RelationshipData" }
          }
        }
      },
      "PaymentRelationships": {
        "type": "object",
        "properties": {
          "payment_submissions": { "$ref": "#/components/schemas/Relationship" },
          "payment_admissions": { "$ref": "#/components/schemas/Relationship" },
          "payment_returns": { "$ref": "#/components/schemas/Relationship" },
          "payment_recalls": { "$ref": "#/components/schemas/Relationship" },
          "payment_reversals": { "$ref": "#/components/schemas/Relationship" }
        }
      },
      "PaymentAttributes": {
        "type": "object",
        "properties": {
          "amount": { "type": "string", "example": "100.00" },
          "currency": { "type": "string", "example": "GBP" },
          "end_to_end_reference": { "type": "string" },
          "numeric_reference": { "type": "string" },
          "payment_scheme": { "type": "string", "default": "FPS" },
          "payment_type": { "type": "string" },
          "processing_date": { "type": "string", "format": "date" },
          "reference": { "type": "string" },
          "scheme_payment_sub_type": { "type": "string" },
          "scheme_payment_type": { "type": "string", "default": "ImmediatePayment" },
          "beneficiary_party": { "$ref": "#/components/schemas/AccountParty" },
          "debtor_party": { "$ref": "#/components/schemas/AccountParty" },
          "charges_information": { "$ref": "#/components/schemas/ChargesInformation" },
          "fx": { "$ref": "#/components/schemas/FxInfo" }
        },
        "required": ["amount", "currency"]
      },
      "PaymentResource": {
        "allOf": [
          { "$ref": "#/components/schemas/Resource" },
          {
            "type": "object",
            "properties": {
              "attributes": { "$ref": "#/components/schemas/PaymentAttributes" },
              "relationships": { "$ref": "#/components/schemas/PaymentRelationships" }
            },
            "required": ["attributes"]
          }
        ]
      },
      "PaymentDataEnvelope": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/PaymentResource" }
        },
        "required": ["data"]
      },
...
      "PaymentSubmissionAttributes": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Lifecycle: accepted -> validation_pending -> limit_check_pending -> limit_check_passed -> released_to_gateway -> queued_for_delivery -> submitted -> delivery_confirmed",
            "enum": ["accepted", "validation_pending", "limit_check_pending", "limit_check_passed", "released_to_gateway", "queued_for_delivery", "submitted", "delivery_confirmed"]
          },
          "submission_date": { "type": "string", "format": "date" },
          "scheme_status_code": { "type": "string" },
          "scheme_transaction_id": { "type": "string" },
          "settlement_cycle": { "type": "integer" },
          "settlement_date": { "type": "string", "format": "date" }
        }
      },
      "PaymentSubmissionResource": {
        "allOf": [
          { "$ref": "#/components/schemas/Resource" },
          {
            "type": "object",
            "properties": {
              "attributes": { "$ref": "#/components/schemas/PaymentSubmissionAttributes" }
            },
            "required": ["attributes"]
          }
        ]
      },
      "PaymentSubmissionDataEnvelope": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/PaymentSubmissionResource" }
        },
        "required": ["data"]
      },
      "PaymentAdmissionAttributes": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Lifecycle: pending -> confirmed",
            "enum": ["pending", "confirmed"]
          },
          "admission_date": { "type": "string", "format": "date" },
          "scheme_transaction_id": { "type": "string" },
          "settlement_cycle": { "type": "integer" },
          "settlement_date": { "type": "string", "format": "date" }
        }
      },
      "PaymentAdmissionResource": {
        "allOf": [
          { "$ref": "#/components/schemas/Resource" },
          {
            "type": "object",
            "properties": {
              "attributes": { "$ref": "#/components/schemas/PaymentAdmissionAttributes" }
            },
            "required": ["attributes"]
          }
        ]
      },
      "PaymentAdmissionDataEnvelope": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/PaymentAdmissionResource" }
        },
        "required": ["data"]
      },
      "AdmissionTaskAttributes": {
        "type": "object",
        "properties": {
          "status": { "type": "string" },
          "assignee": { "type": "string" },
          "name": { "type": "string" }
        }
      },
      "AdmissionTaskResource": {
        "allOf": [
          { "$ref": "#/components/schemas/Resource" },
          {
            "type": "object",
            "properties": {
              "attributes": { "$ref": "#/components/schemas/AdmissionTaskAttributes" }
            },
            "required": ["attributes"]
          }
        ]
      },
      "AdmissionTaskDataEnvelope": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/AdmissionTaskResource" }
        },
        "required": ["data"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "status": { "type": "string" },
                "title": { "type": "string" },
                "detail": { "type": "string" }
              },
              "required": ["status", "title"]
            }
          }
        },
        "required": ["errors"]
      },
...
    }
  }
}